+++ src/components/List/components/public/ListFilterPreset/ListFilterPreset.component.tsx
@@ -7,19 +7,19 @@ import { useListFilterPreset } from "./ListFilterPreset.utils";
 import { Badge, type BadgeProps } from "@/components/Badge";
 import { Text } from "@/components/Text";
 import { useTestId } from "@/utils/props.utils";
+import { normalizeToPaletteColor } from "@utils";
 
 import type {
   ListFilterPresetProps,
   ListFilterPresetStyleProps,
 } from "./ListFilterPreset.types";
+import type { PaletteColor } from "@types";
 
-// TODO: make its background white by default
-// TODO: change the text's color to `${colorKey}-dark` when active
 /**
  * A component that displays a single filter preset button
  * and allows interacting with it.
  *
- * @version 0.0.1
+ * @version 0.0.2
  *
  * @param {ListFilterPresetProps} props
  * @returns {Nullable<JSX.Element>}
@@ -38,6 +38,12 @@ export const ListFilterPreset = (props: ListFilterPresetProps) => {
     [filterPreset]
   );
 
+  const textColor = useMemo<PaletteColor>(() => {
+    if (filterPreset.active)
+      return normalizeToPaletteColor(filterPreset.paletteColor, "dark");
+    return "gray-800";
+  }, [filterPreset.active, filterPreset.paletteColor]);
+
   if (filterPreset.hidden) return null;
 
   return (
@@ -47,7 +53,13 @@ export const ListFilterPreset = (props: ListFilterPresetProps) => {
       {...styleProps}
       data-testid={testId}
     >
-      <Text weight="medium" testId="filter-preset-button-label" fill ellipsis>
+      <Text
+        weight="medium"
+        color={textColor}
+        testId="filter-preset-button-label"
+        fill
+        ellipsis
+      >
         {filterPreset.label}
       </Text>
       <Badge {...badgeProps} />